+++ MatchResult.cs
@@ -11,12 +11,12 @@ namespace AVCapture
     /// </summary>
     class MatchResult
     {
-        public UInt64 EpisodeId;  //Episode ID of identified episode
+        public UInt32 EpisodeId;  //Episode ID of identified episode
         public Int64 SampleTimeDeltaTicks;   //Difference between start of capture file time and start time of matching fingerprints in matched episode
         public UInt32 MatchCount;  //Number of matches at time offset
         public double ConfidenceRatio;  //Confidence ratio
 
-        public MatchResult(UInt64 episodeId, Int64 sampleTimeDeltaTicks, UInt32 matchCount, double confidenceRatio = 0d) {
+        public MatchResult(UInt32 episodeId, Int64 sampleTimeDeltaTicks, UInt32 matchCount, double confidenceRatio = 0d) {
             EpisodeId = episodeId;
             SampleTimeDeltaTicks = sampleTimeDeltaTicks;
             MatchCount = matchCount;
